+++ src/polar_vfs/polar_vfs_interface.c
@@ -789,7 +789,7 @@ vfs_unlink(const char *fname)
 	int			rc = -1;
 	int			kind = -1;
 	const char *vfs_path;
-	vfs_vfd		vfdP;
+	vfs_vfd		vfdP = {0};
 
 	if (fname == NULL)
 		return -1;
@@ -801,6 +801,7 @@ vfs_unlink(const char *fname)
 	vfs_path = polar_vfs_file_type_and_path(fname, &kind);
 
 	vfdP.kind = kind;
+	vfdP.file_name = unconstify(char *, fname);
 	if (polar_vfs_file_before_hook)
 		polar_vfs_file_before_hook(fname, &vfdP, VFS_UNLINK);
 
@@ -819,7 +820,7 @@ vfs_rename(const char *oldfile, const char *newfile)
 	int			kindnew = -1;
 	const char *vfs_old_path;
 	const char *vfs_new_path;
-	vfs_vfd		vfdP;
+	vfs_vfd		vfdP = {0};
 
 	if (oldfile == NULL || newfile == NULL)
 		return -1;
@@ -830,6 +831,7 @@ vfs_rename(const char *oldfile, const char *newfile)
 	vfs_new_path = polar_vfs_file_type_and_path(newfile, &kindnew);
 
 	vfdP.kind = kindold;
+	vfdP.file_name = unconstify(char *, oldfile);
 	if (polar_vfs_file_before_hook)
 	{
 		polar_vfs_file_before_hook(oldfile, &vfdP, VFS_RENAME);
@@ -950,7 +952,7 @@ vfs_truncate(const char *path, off_t len)
 	const char *vfs_path;
 	int			kind = -1;
 	int			rc = -1;
-	vfs_vfd		vfdP;
+	vfs_vfd		vfdP = {0};
 
 	if (path == NULL)
 		return -1;
@@ -962,6 +964,7 @@ vfs_truncate(const char *path, off_t len)
 	vfs_path = polar_vfs_file_type_and_path(path, &kind);
 
 	vfdP.kind = kind;
+	vfdP.file_name = unconstify(char *, path);
 	if (polar_vfs_file_before_hook)
 		polar_vfs_file_before_hook(path, &vfdP, VFS_TRUNCATE);
 
@@ -1081,7 +1084,7 @@ vfs_mkdir(const char *path, mode_t mode)
 	int			rc = -1;
 	int			kind = -1;
 	const char *vfs_path;
-	vfs_vfd		vfdP;
+	vfs_vfd		vfdP = {0};
 
 	if (path == NULL)
 		return -1;
@@ -1089,6 +1092,7 @@ vfs_mkdir(const char *path, mode_t mode)
 	vfs_path = polar_vfs_file_type_and_path(path, &kind);
 
 	vfdP.kind = kind;
+	vfdP.file_name = unconstify(char *, path);
 	if (polar_vfs_file_before_hook)
 		polar_vfs_file_before_hook(path, &vfdP, VFS_MKDIR);
 
@@ -1102,7 +1106,7 @@ vfs_rmdir(const char *path)
 	int			rc = -1;
 	int			kind = -1;
 	const char *vfs_path;
-	vfs_vfd		vfdP;
+	vfs_vfd		vfdP = {0};
 
 	if (path == NULL)
 		return -1;
@@ -1110,6 +1114,7 @@ vfs_rmdir(const char *path)
 	vfs_path = polar_vfs_file_type_and_path(path, &kind);
 
 	vfdP.kind = kind;
+	vfdP.file_name = unconstify(char *, path);
 	if (polar_vfs_file_before_hook)
 		polar_vfs_file_before_hook(path, &vfdP, VFS_RMDIR);